+++ src/memory/aligned_mem.h
@@ -36,4 +36,4 @@ void *aligned_malloc(size_t size, size_t alignment);
 int aligned_free(void *ptr);
 
 // Virtual to physical memory address:
-uintptr_t virt_to_phys_addr(uintptr_t virt_addr, pid_t pid);
+uint32_t virt_to_phys_addr(uint32_t virt_addr, pid_t pid);
